branches main: svlImageFileSource, removed cast on first parameter of `pow`
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@56 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ libs/code/cisstStereoVision/svlImageFileSource.cpp
@@ -278,7 +278,7 @@ int svlImageFileSource::SetSequence(unsigned int numberofdigits, unsigned int fr
 {
     if (numberofdigits > 9 || from > to)
         return SVL_FAIL;
-    unsigned int maxval = static_cast<unsigned int>(pow(static_cast<int>(10), static_cast<int>(numberofdigits)));
+    unsigned int maxval = static_cast<unsigned int>(pow(10.0f, static_cast<int>(numberofdigits)));
     if (from >= maxval || to >= maxval) return SVL_FAIL;
 
     NumberOfDigits = numberofdigits;
